CMakeLists.txt modification for linux compilation with -pthread library

+++ MT-Miner/Profiler.cpp
@@ -1,4 +1,4 @@
 #include "Profiler.h"
 
 bool Profiler::useProfiling = true;
-std::map<std::string, std::chrono::duration<__int64, std::milli>> Profiler::functionDurationMap;
+std::map<std::string, std::chrono::duration<long long, std::milli>> Profiler::functionDurationMap;
